little reddit_parser updates

+++ reddit_pars/load_data.py
@@ -1,5 +1,6 @@
 from bs4 import BeautifulSoup
 import requests
+import os
 url = 'http://files.pushshift.io/reddit/comments/'
 r = requests.get(url)
 soup = BeautifulSoup(r.content)
@@ -12,12 +13,18 @@ import urllib.request
 print('Beginning file download ...')
 k = 0
 for i in list_of_files:
-    dwn = url  + i.strip('./')
-    dwn.replace(' ','_')
-    urllib.request.urlretrieve(dwn, './reddit_data/'+i)
-    k += 1
     print('Downloaded ', k, ' files\nContinue? [yes/no]')
     flag = input()
     if flag == 'no':
         break
+    dwn = url  + i.strip('./')
+    dwn.replace(' ','_')
+    if os.path.exists('./reddit_data/'+i):
+        print('File already exist, download again? [yes/no]')
+        flag = input()
+        if flag == 'no':
+            continue
+    urllib.request.urlretrieve(dwn, './reddit_data/'+i)
+    k += 1
+    
 print('Downloaded files to /reddit_data/')
